Add new systemd-compatible sd-encryptssh hook.
This adds another hook that allows the SSH encryption password
functionality to work with the sd-encrypt hook (and a systemd init).
Unlike the encryptssh hook, the sd-encryptssh hook is an addition to the
sd-encrypt hook and relies on the systemd password agent feature to pass
the entered password on to systemd, which does the actual unlocking.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>